folder renaming, removal of lgoout button, middelware relocation
