+++ src/app/(auth)/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignupForm } from "@/components/forms/SignUpForm";
+import { SignupForm } from "@/components/auth-comps/SignUpForm";
 
 export default function SingUpRoute() {
   return <SignupForm />;
