+++ src/app/(routes)/buy/page.tsx
@@ -1,13 +1,10 @@
 import React from "react";
 
-import { LogoutButton } from "@/components/custom/LogoutButton";
-
 export default function page() {
   return (
     <div>
       {/* Buy page for individuals buying from MTR, this will be feature intensive asf */}
       <h1>Buy</h1>
-      <LogoutButton />
     </div>
   );
 }
