--- src/components/custom/LogoutButton.tsx
@@ -1,12 +1,0 @@
-import { logoutAction } from "@/app/data/actions/auth-actions";
-import { LogOut } from "lucide-react";
-
-export function LogoutButton() {
-  return (
-    <form action={logoutAction}>
-      <button type="submit">
-        <LogOut className="w-6 h-6 hover:text-primary" />
-      </button>
-    </form>
-  );
-}
